Store logical file names in precomp files
We now use logical file names for annotating sources and construct absolute
paths when printing stack traces. This avoids storing absolute paths in
precompiled files which is a problem for packaging modules for Linux
distributions. It also allows for us to add the name of the requested
module to the backtrace output, alleviating the problem with the undeciferable
SHA-1 file names.